feat: reorganize ESA cache scripts and update command references in package.json and documentation
